[Refactor] (test) Move the testing of base classes 'BasePath' and 'BaseNode' to testing module 'model.data'.
